fix(surgery): correct Task-1 test comment + guard parity check against uint8 overflow
Fix 1: the test_locate_overlap_rejects_multi_overlap comment claimed (x,x) is
rejected because x·x commutes, but the weight-3 Steane logical-X anticommutes
with itself (x·x is odd); rejection fires at the overlap-size check (|supp(x)|=3).
Corrected comment only — test body/assertions unchanged.

Fix 2: the anticommutation parity check in _locate_overlap used np.dot on uint8
arrays, which overflows silently for n >= 256. Promote to int64 before dotting so
parity is always correct; behavior on small codes is unchanged.

+++ src/qldpc/circuits/surgery/y_gadget.py
@@ -64,7 +64,7 @@ def _locate_overlap(code: CSSCode, x: np.ndarray, z: np.ndarray) -> int:
     # x and z must anticommute: a logical-X and logical-Z of the SAME logical qubit
     # cross an odd number of times. A commuting pair (even overlap parity) cannot
     # form Ȳ = iX̄Z̄.
-    if int(np.dot(x, z)) % 2 == 0:
+    if int(np.dot(x.astype(np.int64), z.astype(np.int64))) % 2 == 0:
         raise ValueError("x and z commute (x · z is even); they cannot form Ȳ = iX̄Z̄")
 
     overlap = np.where((x.astype(bool)) & (z.astype(bool)))[0]
